Move cache and storage to common package
