+++ extension/ts/common/adapters/storage.ts
