+++ extension/ts/common/util/cache.ts
