+++ extension/ts/content/adapters/book/bookCache.ts
@@ -1,4 +1,4 @@
 import {BookRecord} from "./types";
-import {makeCache} from "../../util/cache";
+import {makeCache} from "../../../common/util/cache";
 const {syncCached, asyncCached, setCache} = makeCache<BookRecord>();
 export {syncCached, asyncCached, setCache};
